+++ core/Print.lua
@@ -50,10 +50,10 @@ function Memento:PrintStatistic()
 
         msg = msg .. Memento_MarkGoldFont(L["statistic.screenshots.encounter.victory"] .. ": ") .. self.dbStatstic.global.events.encounter.victory.count .. " (" .. self.dbStatstic.char.events.encounter.victory.count .. ")\n"
         msg = msg .. Memento_MarkGoldFont(L["statistic.screenshots.encounter.wipe"] .. ": ") .. self.dbStatstic.global.events.encounter.wipe.count .. " (" .. self.dbStatstic.char.events.encounter.wipe.count .. ")\n"
+        msg = msg .. Memento_MarkGoldFont(L["statistic.screenshots.duel"] .. ": ") .. self.dbStatstic.global.events.pvp.duel.count .. " (" .. self.dbStatstic.char.events.pvp.duel.count .. ")\n"
         msg = msg .. Memento_MarkGoldFont(L["statistic.screenshots.levelUp"] .. ": ") .. self.dbStatstic.global.events.levelUp.count .. " (" .. self.dbStatstic.char.events.levelUp.count .. ")\n"
         msg = msg .. Memento_MarkGoldFont(L["statistic.screenshots.death"] .. ": ") .. self.dbStatstic.global.events.death.count .. " (" .. self.dbStatstic.char.events.death.count .. ")\n"
-        msg = msg .. Memento_MarkGoldFont(L["statistic.screenshots.duel"] .. ": ") .. self.dbStatstic.global.events.duel.count .. " (" .. self.dbStatstic.char.events.duel.count .. ")\n"
-        msg = msg .. Memento_MarkGoldFont(L["statistic.screenshots.login"] .. ": ") .. self.dbStatstic.global.events.login.count .. " (" .. self.dbStatstic.char.events.login.count
+        msg = msg .. Memento_MarkGoldFont(L["statistic.screenshots.login"] .. ": ") .. self.dbStatstic.global.events.login.count .. " (" .. self.dbStatstic.char.events.login.count .. ")"
         DEFAULT_CHAT_FRAME:AddMessage(msg)
 	end
 end