+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Memento
-### [2.1.1](https://github.com/diomsg-code/Memento/tree/2.1.1) (2024-10-11)
-[Full Changelog](https://github.com/diomsg-code/Memento/compare/2.1.0...2.1.1) - [Previous Releases](https://github.com/diomsg-code/Memento/releases)
+### [2.2.0-1-alpha](https://github.com/diomsg-code/Memento/tree/2.2.0-1-alpha) (2024-10-17)
+[Full Changelog](https://github.com/diomsg-code/Memento/compare/2.1.1...2.2.0-1-alpha) - [Previous Releases](https://github.com/diomsg-code/Memento/releases)
 
-- Changed: Chat messages for different screenshot events
+- Added: New option category 'Player vs. Player'
+- Added: New event 'Arena completed' - An automatic screenshot can now be taken when you have completed an arena
+- Added: New event 'Battleground completed' - An automatic screenshot can now be taken when you have completed a battleground
 - Updated: deDE, enUS localization
+- Minor code adjustments